Moved validate_required_arg() to common _options.py module
Details:

* The 'pywbemcli instance' command group has a --help-instancename option that
  shows a help message for instance names. The instance commands that have
  this option all define their positional arguments at the Click level with
  required=False, even when they are actually required. If they are required,
  the presence of the argument is validated using the validate_required_arg()
  function that is invoked after the --help-... option has been processed.
  This allows using the --help-... option without having to specify required
  positional arguments.

  This change moves the validate_required_arg() function from the
  pywbemcli/_cmd_instance.py module to the common _options.py module so it
  can be used also by the future pywbemlistener command.
